Rename props -> data_grid_props and move in common folder
